Update project to .Net 7 Improvements to Adamantium.UI

+++ Adamantium/Adamantium.Engine.Graphics/BlendState.cs
@@ -79,7 +79,7 @@ namespace Adamantium.Engine.Graphics
             colorAttachment.DstAlphaBlendFactor = destinationBlend;
             colorAttachment.ColorBlendOp = colorBlendOp;
             colorAttachment.AlphaBlendOp = colorBlendOp;
-            colorAttachment.ColorWriteMask = (uint)colorWriteMask;
+            colorAttachment.ColorWriteMask = colorWriteMask;
             
             return new BlendState(name, state);
         }
